pyproject.toml
[tool.poetry]
name = "sendgrid-cli"
version = "0.1.0"
description = "A command line tool to send emails through SendGrid APIs"
authors = ["Alberto Re <alberto.re@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
sendgrid = "^6.9.7"
click = "^8.1.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "sendgrid_cli.app:sendmail"
